// src/interfaces/spaces.ts
export interface SpaceProps {
  width?: string | undefined;
  height?: string | undefined;

  padding?: string | undefined;
  pt?: string | undefined;
  pr?: string | undefined;
  pb?: string | undefined;
  pl?: string | undefined;

  margin?: string | undefined;
  mt?: string | undefined;
  mr?: string | undefined;
  mb?: string | undefined;
  ml?: string | undefined;
}